Implemented async user veriy dialog
