+++ client/async/UserVerifyBox.js
@@ -1,0 +1,1 @@
+export { default } from '../components/UserVerifyBox'
