+++ client/components/Dialog/Dialog.js
@@ -16,14 +16,20 @@ const dialogComponentNames = {
   [DIALOG_IDS.LOGIN]: 'UserLoginBox',
   [DIALOG_IDS.REGISTER]: 'UserRegisterBox',
   [DIALOG_IDS.RESET_PASSWORD]: 'UserResetPasswordBox',
+  [DIALOG_IDS.VERIFY_USER]: 'UserVerifyBox',
 }
 
 const getDialogForId = (dialogId) => {
+  if (!dialogId) {
+    return null
+  }
+
   const name = dialogComponentNames[dialogId]
   if (!name) {
     console.error(`No dialog component for dialog ${dialogId} found`)
     return null
   }
+
   return <UniversalComponent name={name} />
 }
