+++ client/components/UserVerifyBox/ResendVerificationButton.js
