--- client/pages/UserVerify/Form.js
@@ -1,9 +1,0 @@
-import styled from 'styled-components'
-
-const Form = styled.form`
-  width: 320px;
-  margin-left: auto;
-  margin-right: auto;
-`
-
-export default Form
